+++ index.js
@@ -10,6 +10,12 @@ const FleschKincaid = require('flesch-kincaid');
 
 const {query,validationResult } = require('express-validator');
 const {check} = require( 'express-validator');
+const get = require('./get.js');
+const post = require('./post.js');
+const {insertUser,findUserSafe,rateUser,getUsers} = require('./user.js');
+const {getGroups,joinGroup,insertGroup,findGroup,deleteGroup} = require('./groups.js');
+const {sendPasswordResetEmail} = require('./email.js');
+const {createSession,findSession,allGroupSessions} = require('./session.js');
 
 require('dotenv').config();
 
@@ -40,423 +46,19 @@ app.use(express.static(__dirname + "/www"));
 
 //Eventually we'll need to change this. At the very least, away from the name testDB
 const db_string = "SQL/testDB.db";
-const db = new Database(db_string);
 
-const resetUUIDS = [];
+global.db = new Database(db_string);
+
 
 //joinGroup("test@email.com","Group2");
-
-function getUsers(){
-    return db.prepare("SELECT * FROM Users").all();
-}
+console.log(getUsers());
 
 function analyzeGradeLevel(string){
   return FleschKincaid.grade(string);
 }
 
-function getGroups() {
-  return  db.prepare("SELECT * FROM Groups").all();
-}
-
-function joinGroup(userEmail,groupName){
-  let user = findUserSafe(userEmail);
-  let group = findGroup(undefined,groupName);
-  
-  if(group.length === 1){
-    group = group[0];
-    
-    try{
-      let expr = db.prepare("INSERT INTO GroupMembers (userid,groupid) VALUES(?,?);");
-      let info = expr.run(user.id, group.id);
-      console.log(info);
-      return true;
-    }
-    catch(err){
-      console.log("You're probably trying to add a user to a group that's already in it");
-      console.log(`Error:\n${err}`);
-    }
-  }
-  return false;
-}
-function insertUser(name,email,password){
-  bcrypt
-    .hash(password,10)
-    .then(hash=>{
-      let expr = db.prepare("INSERT INTO Users (name,password,email) VALUES(?,?,?)");
-      let info = expr.run(name,hash,email);
-      console.log(info);
-    })
-    .catch(err=>console.error(err.message));
-}
-
-//Make sure you don't try to bind an object to one of the values.
-//Trust me, it doesn't work
-function insertGroup(owner,name,description){
-  let expr = db.prepare("INSERT INTO Groups (name,owner,description) VALUES(?,?,?)");
-  let info = expr.run(name,findUserSafe(owner.email).id,description);
-  console.log(`Insert group response: ${info}`);
-}
-
-//If you don't need to get the user's password, this is the function you should use
-function findUserSafe(email){
-  let expr = db.prepare("SELECT id,name,email From Users WHERE email=?");
-  let info = expr.get(email);
-  return info;
-}
-
-function findGroup(ownerEmail,name){
-  let user = findUserSafe(ownerEmail);
-  let expr = "";
-  let matchingGroups = undefined;
-  if (user && name){
-    expr = db.prepare("SELECT id,name,description FROM Groups WHERE name=? AND Owner=?");
-    matchingGroups = expr.all(name,user.id);
-  }
-  else if(user){
-    expr = db.prepare("SELECT id,name,description FROM Groups WHERE OWNER=?");
-    matchingGroups = expr.all(user.id);
-  }
-  else if (name){
-    expr = db.prepare("SELECT id,name,description FROM Groups WHERE name=?");
-    matchingGroups = expr.all(name);
-  }
-
-  return matchingGroups;
-}
-
-function rateUser(targetUserId,userId,rating){
-  let target = findUserSafe(targetUserId);
-  let rater = findUserSafe(userId);
-  
-  let expr = "";
-
-  if(target && rater && rating && target.id!==rater.id){
-    try{
-      expr = db.prepare("INSERT INTO UserRatings (rating,ratedby,ratingfor)");
-      let info = expr.run(rating,rater.id,target.id);
-      console.log(info);
-    }
-    catch(err){
-      console.log(`Error: ${err}`);
-    }
-  }
-  else{
-    console.log(`Info: target user: ${target}, rater: ${rater}, rating: ${rating}`);
-  }
-  
-}
-
-function createSession(groupName,time,transcript){
-  let group = findGroup(undefined,groupName);
-  group = group[0];
-  let expr = "";
-  if(!transcript){
-    expr = db.prepare("INSERT INTO Sessions (groupid,time) VALUES (?,?)");
-    expr.run(group.id,time);
-  }
-  else if(transcript){
-    expr = db.prepare("INSERT INTO Sessions (groupid,time,transcript) VALUES (?,?)");
-    expr.run(group.id,time,transcript);
-  }
-}
-function findSession(id){
-  let expr = db.prepare("SELECT * FROM Sessions WHERE id=?");
-  let session = expr.all(id);
-  session = session[0];
-
-  expr = db.prepare("SELECT *,name FROM Sessions INNER JOIN Groups ON Sessions.groupid = Groups.id WHERE Groups.id=?");
-  let name = expr.all(session.groupid);
-  name = name[0].name;
-  console.log(`Found name: ${name}`);
-  session.name=name;
-  return session;
-}
-function allGroupSessions(groupId){
-  let expr = db.prepare("SELECT * FROM Sessions WHERE groupid=?");
-  let results = expr.all(groupId);
-  return results;
-}
-
-function deleteGroup(group){
-  let expr = db.prepare("DELETE FROM Groups where id=?");
-  let info = expr.run(group.id);
-  console.log(info);
-}
-//This creates the object for 
-function getTransporter(){
-  return nodemailer.createTransport({
-    service:'gmail',
-    auth:{
-      user:'ttrpglearning@gmail.com',
-      pass:process.env.pass
-    }
-  });
-}
-function sendMail(transporter,mailOptions){
-  transporter.sendMail(mailOptions,(error,info)=>{
-    if(error){
-      console.log(error);
-    }
-    else{
-      console.log(`Email sent: ${info.response}`);
-    }
-  });
-}
-function sendPasswordResetEmail(email){
-  const transporter = getTransporter();
-  const uuid = crypto.randomUUID();
-  const reset_request = {
-    email:email,
-    uuid:uuid,
-    time:Date.now()
-  }
-  resetUUIDS.push(reset_request);
-  const mailOptions = {
-    from:'ttrpglearning@gmail.com',
-    to:email, //Address to which you want to send
-    subject:'Password Reset', //subject
-    html:`Click <a href="http://localhost:8000/recover/${uuid}">here</a> to reset your password` //body of email
-  };
-  sendMail(transporter,mailOptions);
-}
-
-
-app.get("/",(req,res)=>{
-  
-  if(req.session.username){
-    let user = getUsers().find((user)=>user.email === req.session.username);
-    res.render("index",{user:user});
-  }
-  else{
-    res.render("index");
-  }
-});
-app.get("/recover/:id/",(req,res)=>{
-  let id = req.params.id;
-  let validRequest = resetUUIDS.find(u=>u.uuid === id);
-  let index = undefined;
-  if(validRequest){
-    index = resetUUIDS.findIndex(u=>u.uuid === id);
-  }
-  else{
-    res.status(400).send("Invalid link");
-  }
-  console.log(validRequest);
-  console.log("Index is: "+index);
-  if(validRequest && Date.now() - validRequest.time <= 9000000){
-    console.log("Allow password resetting");
-    console.log(Date.now() - validRequest.time);
-    //res.render("newpassword");
-    res.send("Worked!");
-  }
-  else{
-    console.log("Password reset no longer valid");
-    resetUUIDS = resetUUIDS.filter(u=>u.uuid !== id);
-    res.status(401).send("Link no longer valid");
-  }
-  
-});
-app.get("/group",(req,res)=>{
-  let users = getUsers();
-  res.render("creategroup",{users:users});
-});
-app.get("/check",(req,res)=>{
-  if(req.session.username){
-    res.send(`You're logged in as ${req.session.username}`);
-  }
-  else{
-    res.send("Please log in");
-  }
-});
-app.get("/register",(req,res)=>{
-  res.render("registration");
-});
-app.get("/login",(req,res)=>{
-  res.render("login");
-});
-app.get("/recovery",(req,res)=>{
-  res.render("recovery");
-});
-app.get("/userpage",(req,res)=>{
-  if(req.session.loggedIn){
-    let user = getUsers().find((user)=> user.email === req.session.username);
-    res.render("userpage",{user:user});
-  }
-  else{
-    res.redirect("/login");
-  }
-});
-app.get("/session/:id/",(req,res)=>{
-  let id = req.params.id;
-  
-  let session = findSession(id);
-  let groups = getGroups();
-  console.log(session);
-  session.time = session.time.slice(0,session.time.length-3);
-  
-  if(session != undefined){
-    console.log(session);
-    res.render("sessionpage",{session:session,groups:groups});
-  }
-  else{
-    //This should be changed to return an error? Or redirect back to where they came from?
-    res.render("sessionpage",{groups:groups});
-  }
-});
-app.post("/recoverpassword",[check("password","You must supply a password").notEmpty(),check("confirmpassword","You must confirm your password").notEmpty()],(req,res)=>{
-  const errors = validationResult(req);
-  let localErrors = [];
-  let password = req.body.password;
-  let confirmpassword = req.body.password;
-
-  if(password !== confirmpassword){
-    localErrors.push("Your passwords do not match");
-  }
-  if(errors.length>0 || localErrors.length > 0){
-    //{errors:errors.array()}
-    res.render("whatever the password recovery/change page is");
-  }
-  else{
-    //Do the thing
-  }
-  
-  
-});
-//This needs some checking, probably
-app.post("/session",(req,res)=>{
-  let time = req.body.time;
-  let group = req.body.group;
-  time = time.replace("T"," ")+":00";
-  createSession(group,time);
-  console.log(time);
-  console.log(group);
-  res.render("sessionpage");
-});
-
-app.post("/recovery",[check('email',"Please enter a valid email").isEmail()],(req,res)=>{
-  const errors = validationResult(req);
-
-  if(!errors.isEmpty()){
-    res.render("recovery");
-  }
-  else{
-    
-    let email = req.body.email;
- 
-    sendPasswordResetEmail(email);
-    res.send("A password reset email has been sent <FiX THIS LATER>");
-  }
-});
-
-app.post("/register",[check('name',"Please enter a name").notEmpty(),check('email',"Please enter a valid email").isEmail(),check('password',"Please enter a password").notEmpty(),check('confirmPassword',"Please confirm your password").notEmpty()],(req,res)=>{
-  
-  const errors = validationResult(req);
-  console.log(errors);
-  if(!errors.isEmpty()){
-    return res.render("registration",{errors:errors.array()});
-  }
-  let name = req.body.name;
-  let email = req.body.email;
-  let password = req.body.password;
-
-  insertUser(name,email,password);
-  //res.send("Pretend that we definitely did that right. Eventually we'll check this server-side");
-  res.render("registration",{message:"User account created!"});
-});
-
-app.post("/login",[
-  query('email').isEmail()], (req, res) => {
-    const users = getUsers();
-    const email = req.body.email;
-    const password = req.body.password;
-    const user = users.find((user) => user.email === email);
-
-    if (!user) {
-      console.log("User not found");
-      return res.render("login");
-    }
-
-    bcrypt.compare(password, user.password)
-      .then((result) => {
-	if (result) {
-          req.session.loggedIn = true;
-          req.session.username = user.email;
-          req.session.role = "user";
-          console.log("User login correct");
-          res.render("login",{message:"You've logged in successfully!"});
-	} else {
-          console.log("Password is wrong");
-          res.render("login",{message:"Incorrect login info"});
-	}
-      })
-      .catch((err) => {
-	console.error(`Error with bcrypt: ${err}`);
-	res.render("login");
-      });
-});
-app.post("/logout",(req,res)=>{
-  if(req.session.username){
-    console.log(`User ${req.session.username} has been logged out`);
-  }
-  else{
-    console.log("No one is currently logged in");
-  }
-  req.session.loggedIn = undefined;
-  req.session.username = undefined;
-  req.session.role = undefined;
-  res.render("index");
-});
-
-app.post("/group",(req,res)=>{
-  let email = req.body.owner;
-  let name = req.body.name;
-  let description = req.body.description;
-  let users = getUsers();
-  console.log(`Email:${email}`);
-  console.log(`Name:${name}`);
-  const user = users.find((user)=> user.email === email);
-  if(!user){
-    return res.render("creategroup",{error:"User doesn't exist"});
-  }
-  insertGroup(user,name,description);
-  res.render("creategroup",{message:"Successfull"});
-});
-app.post("/findgroup",(req,res)=>{
-  let name = req.body.name;
-  //TODO join Groups and Users table
-  let owner = req.body.owneremail;
-  let groups = getGroups();
-  let group = "";
-  if(name){
-    group = groups.find((group)=> group.name === name);
-    console.log(group);
-  }
-
-});
-app.post("/deletegroup",(req,res)=>{
-  let groups = getGroups();
-  let groupName = req.body.name;
-  let group = groups.find((group)=> group.name === groupName);
-  if(!group){
-    return res.render("group",{message:"No such group exists"});
-  }
-  deleteGroup(group);
-});
-app.post("/groupsearch",(req,res)=>{
-  
-  let ownerEmail = req.body.email;
-  let groupName = req.body.groupName;
-  let matches = findGroup(ownerEmail,groupName);
-  if(matches){
-    res.status(200).json(matches);
-  }
-  else if (!ownerEmail && !groupName){
-    res.status(400).send("Bad request");
-  }
-  else{
-    res.status(200).send("No matches");
-  }
-});
+app.use('/',get);
+app.use('/',post);
 
 app.listen(port,()=>{
   console.log(`Running server at http://localhost:8000`);
